# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Wedding Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /weddings/{weddingId}/coordinators:
    get:
      operationId: listCoordinators
      tags: [coordinators]
      summary: List coordinators
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Coordinator list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coordinator"
    post:
      operationId: createCoordinator
      tags: [coordinators]
      summary: Create coordinator
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoordinatorInput"
      responses:
        "201":
          description: Coordinator created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coordinator"
  /weddings/{weddingId}/coordinators/{id}:
    patch:
      operationId: updateCoordinator
      tags: [coordinators]
      summary: Update coordinator
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoordinatorInput"
      responses:
        "200":
          description: Coordinator updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coordinator"
    delete:
      operationId: deleteCoordinator
      tags: [coordinators]
      summary: Delete coordinator
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /weddings/{weddingId}/budget-categories:
    get:
      operationId: listBudgetCategories
      tags: [budget]
      summary: List budget categories
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Budget categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BudgetCategory"
    post:
      operationId: createBudgetCategory
      tags: [budget]
      summary: Create budget category
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetCategoryInput"
      responses:
        "201":
          description: Budget category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetCategory"
  /weddings/{weddingId}/budget-categories/{id}:
    patch:
      operationId: updateBudgetCategory
      tags: [budget]
      summary: Update budget category
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetCategoryInput"
      responses:
        "200":
          description: Budget category updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetCategory"
    delete:
      operationId: deleteBudgetCategory
      tags: [budget]
      summary: Delete budget category
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /weddings/{weddingId}/budget-items:
    get:
      operationId: listBudgetItems
      tags: [budget]
      summary: List budget items
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
        - name: categoryId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Budget items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BudgetItem"
    post:
      operationId: createBudgetItem
      tags: [budget]
      summary: Create budget item
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetItemInput"
      responses:
        "201":
          description: Budget item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetItem"
  /weddings/{weddingId}/budget-items/{id}:
    patch:
      operationId: updateBudgetItem
      tags: [budget]
      summary: Update budget item
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetItemInput"
      responses:
        "200":
          description: Budget item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetItem"
    delete:
      operationId: deleteBudgetItem
      tags: [budget]
      summary: Delete budget item
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /weddings/{weddingId}/budget-summary:
    get:
      operationId: getBudgetSummary
      tags: [budget]
      summary: Get budget summary
      parameters:
        - name: weddingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Budget summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetSummary"
...
components:
  schemas:
...
    RegisterInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, planner, coordinator, couple, guest]
      required:
        - name
        - email
        - password
        - role
    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, planner, coordinator, couple, guest]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - createdAt
...
    Coordinator:
      type: object
      properties:
        id:
          type: integer
        weddingId:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        role:
          type: string
        permissions:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - weddingId
        - name
        - role
        - permissions
        - createdAt
    CoordinatorInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        role:
          type: string
        permissions:
          type: array
          items:
            type: string
...
    BudgetCategory:
      type: object
      properties:
        id:
          type: integer
        weddingId:
          type: integer
        name:
          type: string
        estimatedTotal:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - weddingId
        - name
        - estimatedTotal
        - createdAt
    BudgetCategoryInput:
      type: object
      properties:
        name:
          type: string
        estimatedTotal:
          type: number
    BudgetItem:
      type: object
      properties:
        id:
          type: integer
        weddingId:
          type: integer
        categoryId:
          type: integer
        name:
          type: string
        estimatedCost:
          type: number
        actualCost:
          type: ["number", "null"]
        vendor:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isPaid:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - weddingId
        - categoryId
        - name
        - estimatedCost
        - isPaid
        - createdAt
    BudgetItemInput:
      type: object
      properties:
        categoryId:
          type: integer
        name:
          type: string
        estimatedCost:
          type: number
        actualCost:
          type: ["number", "null"]
        vendor:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isPaid:
          type: boolean
    BudgetSummary:
      type: object
      properties:
        totalEstimated:
          type: number
        totalActual:
          type: number
        totalPaid:
          type: number
        totalRemaining:
          type: number
        byCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategoryBudgetSummary"
      required:
        - totalEstimated
        - totalActual
        - totalPaid
        - totalRemaining
        - byCategory
    CategoryBudgetSummary:
      type: object
      properties:
        categoryId:
          type: integer
        categoryName:
          type: string
        estimated:
          type: number
        actual:
          type: number
        paid:
          type: number
      required:
        - categoryId
        - categoryName
        - estimated
        - actual
        - paid
...