# lib/api-zod/src/generated/types/budgetCategory.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wedding Management API
 * OpenAPI spec version: 0.1.0
 */

export interface BudgetCategory {
  id: number;
  weddingId: number;
  name: string;
  estimatedTotal: number;
  createdAt: Date;
}